completed health hud
Health hud now apropriate changes based on the amount of current health.
However, there is a bug where the player is not taking damage from the
enemies. To test you must manually change current health while game is
running.

If someone has a chance to look at why the Enemy Attack script is not
working correctly I would appreciate it. I looked at it and can't figure
it out. I think it may be something to do with the player tag.